+++ start.py
@@ -1,0 +1,4 @@
+import subprocess
+
+# Execute the Python file
+subprocess.call(["python", "module/chatbot/chatbot.py"])
